Update
In the new version, the 'else' instructions have been added, which if the cell in column H is empty, it takes a distinctive image not related to the questions. The reason for adding these instructions is due to the fact that in the Unity engine, the images were numbered in a hierarchical manner and referring to a particular image was problematic. Now, the only problem is the amount of images loaded into the arrays at the start of the game. Before running the script, the paths must be changed if they do not match and retrieve the question sheets.